Change last uint

+++ sinabs/backend/dynapcnn/chip_factory.py
@@ -89,7 +89,7 @@ class ChipFactory:
             ev.x = row[3]
             ev.y = row[2]
             ev.feature = row[1]
-            ev.timestamp = np.uint64( ( row[0].item() * 1e6 * dt ) + ( delay_factor * 1e6 ) )  # Time in uS
+            ev.timestamp = int( ( row[0].item() * 1e6 * dt ) + ( delay_factor * 1e6 ) )  # Time in uS
             events.append(ev)
         return events
